Added Debug and Error function
Debug function takes if you're debugging or not, and an io writer. It
then returns a function to be used like the other print functions. I
have found this to be useful when using multi io writer.

Error function also takes an io writer. Again I found this useful when
using multi io writers.